Trait ultrustar::core::platform::PlatformApi

pub trait PlatformApi: Sized {
    type Settings;
    type Renderer: Renderer;
    type InitError;
    type GlWindow: Sized;
    type Audio: PlatformApi;
    fn create_gl_window(&self) -> Result<Self::GlWindow, Error>;

    fn load_userdata(id: &str) -> Result<UserData, Error>;

    fn persist_userdata(data: &UserData) -> Result<(), Error>;

    fn init(settings: Self::Settings) -> Result<Self, Self::InitError>;

    fn run<F>(self, main_loop: F)
    where
        F: 'static + FnMut(&Event<'_, Signals>, &EventLoopWindowTarget<Signals>);

    fn create_renderer(
        &self, 
        settings: &<Self::Renderer as Renderer>::InitSettings
    ) -> Result<Self::Renderer, Error>;
}

Documentation trait for what APIs a platform implementation needs to expose

Associated Types

type Settings

Platform-specific initialization settings

type Renderer: Renderer

The (possibly polymorphic) type of renderer to be used

type InitError

Representation of errors that may occur during initialization

type GlWindow: Sized

Implementation of a window offering OpenGl functionality on the current platform.

type Audio: PlatformApi

Low-level audio interfaces.

Required methods

fn create_gl_window(&self) -> Result<Self::GlWindow, Error>

Create a new window which offers OpenGl graphics capabilities

Errors

Return an error if the creation of the gl-capable window failed.

fn load_userdata(id: &str) -> Result<UserData, Error>

Load user data from persistent storage

Errors

In case user data cannot be loaded and requested user is not “default”, either.

fn persist_userdata(data: &UserData) -> Result<(), Error>

Store user data to persistent storage

Errors

In case I/O failed

fn init(settings: Self::Settings) -> Result<Self, Self::InitError>

Initializes (instantiates) the platform

Errors

Platform-specific initialization errors

fn run<F>(self, main_loop: F) where
F: 'static + FnMut(&Event<'_, Signals>, &EventLoopWindowTarget<Signals>),

Allow hooking into the event loop

fn create_renderer(
&self,
settings: &<Self::Renderer as Renderer>::InitSettings
) -> Result<Self::Renderer, Error>

Create a renderer instance

Errors

Forwards any errors of the underlying graphics subsystem which prevent successful instantiation.